Writing identified mutex locked traces is conceptually flawed
Traces are conceptually hierarchical: a trace that begins after another trace must end before that trace. But mutex locks are not hierarchical.

For now, remove tracing of mutex locks. I think to report this information, we need to have a separate track for the mutex, which is a different kind of track than a thread.